+++ src/ChildValidatorsAwareTrait.php
@@ -2,13 +2,12 @@
 
 namespace Dhii\Validation;
 
+use stdClass;
 use Traversable;
-use Exception as RootException;
 use InvalidArgumentException;
-use Dhii\Util\String\StringableInterface as Stringable;
 
 /**
- * Functionality for retrieving child validators.
+ * Awareness of child validators.
  *
  * @since [*next-version*]
  */
@@ -19,16 +18,16 @@ trait ChildValidatorsAwareTrait
      *
      * @since [*next-version*]
      *
-     * @var array|Traversable|null
+     * @var array|Traversable|stdClass
      */
     protected $childValidators;
 
     /**
-     * Retrieves the spec associated with this instance.
+     * Retrieves the list of child validators associated with this instance.
      *
      * @since [*next-version*]
      *
-     * @return array|Traversable|null The spec.
+     * @return array|Traversable|stdClass The list of validators.
      */
     protected function _getChildValidators()
     {
@@ -40,39 +39,35 @@ trait ChildValidatorsAwareTrait
     }
 
     /**
-     * Assigns a spec to this instance.
+     * Retrieves the child validators.
      *
      * @since [*next-version*]
      *
-     * @param array|Traversable|null $spec The spec.
+     * @param ValidatorInterface[]|Traversable|stdClass $validators A list of validators.
+     *
+     * @throws InvalidArgumentException If the validators list is invalid.
      */
-    protected function _setChildValidators($spec)
+    protected function _setChildValidators($validators)
     {
-        if (!is_null($spec) && !is_array($spec) && !($spec instanceof Traversable)) {
-            throw $this->_createInvalidArgumentException($this->__('Invalid specification'), null, null, $spec);
-        }
+        $validators = $this->_normalizeIterable($validators);
 
-        $this->childValidators = $spec;
+        $this->childValidators = $validators;
     }
 
     /**
-     * Creates a new Dhii invalid argument exception.
+     * Normalizes an iterable.
+     *
+     * Makes sure that the return value can be iterated over.
      *
      * @since [*next-version*]
      *
-     * @param string|Stringable|null $message  The error message, if any.
-     * @param int|null               $code     The error code, if any.
-     * @param RootException|null     $previous The inner exception for chaining, if any.
-     * @param mixed|null             $argument The invalid argument, if any.
+     * @param mixed $iterable The iterable to normalize.
      *
-     * @return InvalidArgumentException The new exception.
+     * @throws InvalidArgumentException If the iterable could not be normalized.
+     *
+     * @return array|Traversable|stdClass The normalized iterable.
      */
-    abstract protected function _createInvalidArgumentException(
-            $message = null,
-            $code = null,
-            RootException $previous = null,
-            $argument = null
-    );
+    abstract protected function _normalizeIterable($iterable);
 
     /**
      * Translates a string, and replaces placeholders.